+++ apps/src/componentLibrary/card/card.story.tsx
@@ -9,8 +9,7 @@ import i18n from '@cdo/locale';
 
 import {Card} from './Card';
 
-import cardStyles from '@cdo/apps/componentLibrary/card/Card/card.module.scss';
-import styles from '@cdo/apps/lib/ui/lti/link/LtiLinkAccountPage/link-account.module.scss';
+import styles from '@cdo/apps/lib/ui/simpleSignUp/link-account.module.scss';
 
 const meta: Meta<typeof Card> = {
   title: 'DesignSystem/Card', // eslint-disable-line storybook/no-title-property-in-meta
@@ -24,7 +23,7 @@ export const UserSignupCard: Story = {
   render: () => (
     <Card data-testid={'existing-account-card'}>
       <CardHeader title={i18n.ltiLinkAccountExistingAccountCardHeaderLabel()} />
-      <CardContent className={cardStyles.cardContent}>
+      <CardContent className={styles.cardContent}>
         {i18n.ltiLinkAccountExistingAccountCardContent({
           providerName: 'Canvas',
         })}
