+++ apps/src/lib/ui/simpleSignUp/lti/feedback/LtiFeedbackBanner.tsx
@@ -9,7 +9,7 @@ import {
 } from '@cdo/apps/redux/lti/ltiFeedbackReducer';
 import i18n from '@cdo/locale';
 
-import FeedbackBanner, {BANNER_STATUS} from '../../feedback/FeedbackBanner';
+import FeedbackBanner, {BANNER_STATUS} from '../../../feedback/FeedbackBanner';
 
 /**
  * LtiFeedbackBanner component
