+++ apps/src/lib/ui/simpleSignUp/lti/link/LtiLinkAccountPage/cards/ExistingAccountCard/index.tsx
@@ -6,7 +6,7 @@ import {
 } from '@cdo/apps/componentLibrary/card';
 import FontAwesomeV6Icon from '@cdo/apps/componentLibrary/fontAwesomeV6Icon';
 import classNames from 'classnames';
-import styles from '@cdo/apps/lib/ui/lti/link/LtiLinkAccountPage/link-account.module.scss';
+import styles from '../../../../../link-account.module.scss';
 import {buttonColors, Button} from '@cdo/apps/componentLibrary/button';
 import React, {useContext} from 'react';
 import i18n from '@cdo/locale';
@@ -62,7 +62,7 @@ const ExistingAccountCard = () => {
       </CardContent>
       <CardActions>
         <Button
-          className={styles.button}
+          className={classNames(styles.button, styles.cardSecondaryButton)}
           color={buttonColors.purple}
           type={'primary'}
           size="m"
