+++ apps/src/lib/ui/simpleSignUp/lti/registration/styles.module.scss
