+++ apps/src/lib/ui/simpleSignUp/lti/upgrade/LtiUpgradeAccountDialog.story.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {StoryFn} from '@storybook/react';
-import LtiUpgradeAccountDialog from '@cdo/apps/lib/ui/lti/upgrade/LtiUpgradeAccountDialog';
+import LtiUpgradeAccountDialog from '@cdo/apps/lib/ui/simpleSignUp/lti/upgrade/LtiUpgradeAccountDialog';
 
 export default {
   component: LtiUpgradeAccountDialog,
