+++ apps/src/sites/studio/pages/lti/v1/dynamic_registration.js
@@ -6,7 +6,7 @@
  */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import LtiDynamicRegistrationPage from '@cdo/apps/lib/ui/lti/registration/LtiDynamicRegistrationPage';
+import LtiDynamicRegistrationPage from '@cdo/apps/lib/ui/simpleSignUp/lti/registration/LtiDynamicRegistrationPage';
 import getScriptData from '@cdo/apps/util/getScriptData';
 
 document.addEventListener('DOMContentLoaded', () => {
