+++ apps/src/sites/studio/pages/lti/v1/iframe.js
@@ -6,7 +6,7 @@
  */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import LtiIframePage from '@cdo/apps/lib/ui/lti/iframe/LtiIframePage';
+import LtiIframePage from '@cdo/apps/lib/ui/simpleSignUp/lti/iframe/LtiIframePage';
 import getScriptData from '@cdo/apps/util/getScriptData';
 
 document.addEventListener('DOMContentLoaded', () => {
