+++ apps/src/sites/studio/pages/lti/v1/upgrade_account.js
@@ -7,7 +7,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import getScriptData from '@cdo/apps/util/getScriptData';
-import LtiUpgradeAccountDialog from '@cdo/apps/lib/ui/lti/upgrade/LtiUpgradeAccountDialog';
+import LtiUpgradeAccountDialog from '@cdo/apps/lib/ui/simpleSignUp/lti/upgrade/LtiUpgradeAccountDialog';
 
 document.addEventListener('DOMContentLoaded', () => {
   const scriptData = getScriptData('ltiUpgradeAccountFormData');
